requirements.txt
PyQt6>=6.5.0
hidapi>=0.14.0
tomli>=2.0.1; python_version < "3.11"
tomli-w>=1.0.0
